code/requirements.txt
# Virtual Users Research - Requirements
# Python 3.9+

# Core scientific computing
numpy>=1.21.0
scipy>=1.7.0
scikit-learn>=1.0.0

# Configuration and data handling
PyYAML>=6.0
pandas>=1.3.0

# HTTP requests for data downloading
requests>=2.25.0

# Progress tracking and logging
tqdm>=4.62.0

# Optional: For enhanced data analysis (if needed)
matplotlib>=3.4.0  # For result visualization
seaborn>=0.11.0    # For statistical visualization

# Development dependencies
pytest>=6.2.0      # For testing
black>=21.0.0      # Code formatting
flake8>=3.9.0      # Linting